finished all i2c interfaces in Roller485 class

+++ src/drivers/Roller485/src/Roller485.h
@@ -33,17 +33,17 @@ class Roller485
 	void setMode(motor_mode_t mode);
 	void setRGBUserMode();
 	void setRGBColor(uint8_t red, uint8_t green, uint8_t blue);
-	void setMaxCurrent(int32_t max_current, motor_mode_t mode);
-	void setSpeed(int32_t speed);
-	int32_t getSpeed();
+	void setMaxCurrent(float max_current, motor_mode_t mode);
+	void setSpeed(float speed);
+	float getSpeed();
 	void setSpeedPID(float p, float i, float d);
 	void setPosition(float position);
 	float getPosition();
 	void setPositionPID(float p, float i, float d);
-	void setCurrent(int32_t current);
-	int32_t getCurrent();
-	int32_t getInternalTemperature();
-	int32_t getEncoderCounter();
+	void setCurrent(float current);
+	float getCurrent();
+	float getInternalTemperature();
+	float getEncoderCounter();
 
       private:
 	int sendAndCheck(uint8_t *data, size_t length);
